7.3. Criação da tela base de projeto e envio do objeto selecionado
