+++ src/main/webapp/restricted/projeto.xhtml
@@ -1,0 +1,30 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml"
+	xmlns:f="http://xmlns.jcp.org/jsf/core"
+	xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+	xmlns:h="http://xmlns.jcp.org/jsf/html"
+	xmlns:p="http://primefaces.org/ui">
+
+<h:head>
+	<title>Projeto</title>
+</h:head>
+
+<f:metadata>
+	<f:viewAction action="#{projetoMB.init()}"/>
+</f:metadata>
+
+<h:body>
+	<h:form id="f-projeto">
+		<p:growl id="messages">
+			<p:autoUpdate />
+		</p:growl>
+
+		<h:panelGroup id="pg-projeto" layout="block"
+			style="width: 1000px; height: 750px; margin: 50px auto 20px auto; border-radius: 20px; border-style: solid; border-width: 2px; border-color: #bdc4ca; box-shadow: 2px 6px 10px #bdc4ca;">
+			<h:outputLabel id="title-dashboard" value="#{projetoMB.projeto.nome}"
+				style="font-size: 30px; background-color: #008cff; color: white; padding: 15px 35px 5px 10px; margin: 15px 5px 5px 5px; border-radius: 10px; font-family: Arial, sans-serif;" />
+
+		</h:panelGroup>
+	</h:form>
+</h:body>
+</html>
